Finish lifting the per-kind node methods out of SceneDocument
PR #409 moved the two largest per-kind method groups (code_review,
gitlink) into their own mixins and left the other thirteen kinds behind.
This moves those thirteen, which is the rest of that job: 42 members and
983 lines out of graph.py, taking it from 2,591 lines to 1,502.

Five new mixins, grouped by what the kinds actually have in common
rather than one module per kind:

  nodes_agent_runs.py    AgentRunOps        web_research, artifact,
                                            code_sandbox - the three kinds
                                            with a start/progress/complete/
                                            fail run lifecycle
  nodes_planning.py      PlanningOps        plan, harness
  nodes_conversational.py ConversationalOps chat, conversation
  nodes_content.py       ContentOps         document, thinking, html, note
  nodes_visual.py        VisualOps          chart, image

Every one of the 42 moved members was verified AST-identical to its
pre-move version (ast.unparse round-trip against
`git show HEAD:backend/domain/graph.py`), verified absent from
SceneDocument's own body afterwards, verified reachable on SceneDocument,
and verified to be defined exactly once across the whole MRO. Methods are
regrouped by kind inside each new module instead of keeping the order that
successive increments happened to append them in; nothing else changed.

Two supporting edits:

SceneDocumentParts gains `adopt_pending_system_prompt`, which add_chat_node
calls and which stays on SceneDocument, and its `add_chat_node` declaration
is tightened from a `*args: Any, **kwargs: Any` hedge to the real
signature. The hedge was fine while the body lived in SceneDocument itself;
now that ConversationalOps defines it, an inexact declaration in a base
class is an incompatible-override error rather than a harmless fiction.

#409 also left two section-header comments in graph.py describing methods
it had just moved away, plus 21 stray blank lines where they used to sit.
The headers carried real information about those kinds' import posture, so
that text moves into the nodes_gitlink.py and nodes_code_review.py module
docstrings rather than being dropped.

Test plan: full suite, 3,226 passed / 19 skipped. ruff clean.
